few more options in tests

+++ benchmarks/mou_ppd_fitness_tests.py
@@ -73,9 +73,11 @@ def ppd_part1_test():
 def ppd_fitness(ppd_beta = 0.5, master_dir = "master",test_d="ppd_fitness_test_sklearn"):
 
     t_d = os.path.join(test_d, "template")
-    m_d = os.path.join(test_d, master_dir)
+    m_d = None
+    if master_dir is not None:
+        m_d = os.path.join(test_d, master_dir)
 
-    if os.path.exists(m_d):
+    if m_d is not None and os.path.exists(m_d):
         shutil.rmtree(m_d)
     pst = pyemu.Pst(os.path.join(t_d, "pest.pst"))
     obs = pst.observation_data
@@ -214,7 +216,10 @@ def ppd_part_2_3_test():
     
 
 if __name__ == "__main__":
-    #ppd_part1_test()
+    #test_d = "ppd_fitness_test_sklearn"
+    #parsum1 = ppd_fitness(ppd_beta = 0.5, master_dir = None,test_d=test_d)
+    #ppd_part2()
+    ppd_part1_test()
     ppd_part_2_3_test()
